requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
pydantic>=2.0.0
pyyaml>=6.0
httpx>=0.27.0
aiofiles>=24.0.0
supervisor>=4.2.0
